solver for multi-asset CFMM (#1429)
Closes: #1369

## What is the purpose of the change

This PR implements the solver for our multi-asset CFMM (as outlined in the previous PR for issue #1369).

This is an expansion of the original two-asset CFMM we extended from Solidly, which we expand here to accommodate n-asset pools without significantly increasing complexity or runtime. While this type of multi-asset pool was one that we found was perhaps unnecessarily complex for volatile assets, it is critical from a capital efficiency standpoint that we have the option for Curve-like 3pools and 4pools to not fragment stablecoin liquidity (and therefore also LP incentives) within our stableswap, especially as we compete for stablecoin liquidity with many other stableswaps.

Our final multi-asset CFMM is `xyz(x^2 + y^2 + w) = k`, where z represents the product of the reserves of all assets that are not being swapped between, and w represents the sum of all of their squares. This PR implements a solver that finds the correct output for of a requested token `a` given an input of another token `b` in the pool and accommodates swaps between any two assets `x` and `y` in an n-asset stableswap. The polynomial we solve for can be described as:

`(x - a)(y + b)z((x - a)^2 + (y +b)^2 + w)`, where we solve for `a`. The solution is quite complex and not easy to visually parse without the abstractions we have laid out in the code comments, so please refer to those for more comprehensive and readable implementation details.

## Brief change log

  - Implemented `solveCfmmMulti` function in `x/gamm/pool-models/stableswap/amm.go` using the derivations set up in code comments
  - Implemented binary search logic for this new multi-asset CFMM
  - Set up tests for multi-asset pools and expanded existing two-asset stableswap tests to also check consistency between our two CFMM implementations
  - Cleaned up/simplified the math for the solver to make it more readable and simpler to compute (e.g. limiting the max bitlen to avoid risking integer overflows)

## Testing and Verifying

The tests for this change line up in coverage with those for our two-asset stableswap CFMM and can be found in `amm_test.go`.

## Documentation and Release Note

  - Does this pull request introduce a new feature or user-facing behavior changes? (yes)
  - Is a relevant changelog entry added to the `Unreleased` section in `CHANGELOG.md`? (no)
  - How is the feature or change documented? (not documented)