minor: fix comments about pool state update in `CalcJoinPoolShares` (#2344)
## What is the purpose of the change

`CalcJoinPoolShares` has a misleading/incorrect comment that claims that `updateIntermediaryPoolAssetsLiquidity` updates "effective pool state". 

This might lead someone to believe that `CalcJoinPoolShares` mutates pool state, which is not true and can lead to a lot of confusion on how our balancer pool joins work. This isn't immediately apparent from the code unless you're already familiar with other parts of the gamm module, so I figured it would be worth correcting in a small PR.

## Brief Changelog

- Minor comment change

## Testing and Verifying

This change is a trivial rework / code cleanup without any test coverage.

## Documentation and Release Note

  - Does this pull request introduce a new feature or user-facing behavior changes? (no)
  - Is a relevant changelog entry added to the `Unreleased` section in `CHANGELOG.md`? (no)
  - How is the feature or change documented? (not applicable)