chore: godoc pointing to wrong version (#2559)
## What is the purpose of the change

Our godoc and go report card were both pointing to the latest v1 branch, when in reality it should be pointing to the v11 package which then pulls the latest version from that.

Lastly, this PR gets rid of a broken / nonexistent link.


## Testing and Verifying

This change is a trivial rework / code cleanup without any test coverage

## Documentation and Release Note

  - Does this pull request introduce a new feature or user-facing behavior changes? no
  - Is a relevant changelog entry added to the `Unreleased` section in `CHANGELOG.md`? no
  - How is the feature or change documented? not applicable